requirements.txt
altair==4.2.0
asgiref==3.5.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0

charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
Django==4.1.1
entrypoints==0.4
fortran-language-server==1.12.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
image==1.5.33
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pickle-mixin==1.0.2
Pillow==9.2.0
protobuf==3.20.2
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.1
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sqlparse==0.4.2
streamlit==1.12.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wincertstore==0.2
zipp==3.8.1
gunicorn